Create Dog.py
Create a Dog class. A Dog should have two instance variables — a name and an age.

Write a constructor that sets a dog’s name equal to the second value passed to the constructor and the age to the third value. As a reminder, the first value should be self.

Finally, a Dog should have two methods:

introduce(). This method should return the String "Hello, my name is <name> and I am <age> years old". Make sure to cast age as a String in order to concatinate it with the rest of your sentence.
is_older(). This method should take a Dog as a parameter. This method should return True if this Dog‘s age is greater than or equal to the age of the Dog passed as a parameter. If the other Dog is older, this method should return False.
Create two dogs and test your two methods.